2단계 - 리뷰 반영
